Distribute/package/install docs, scripts, and systemd unit files.
Add installation of systemd unit files to shifter.spec.  This can be
disabled by passing `--without systemd` to `rpmbuild`.  Install with
the `shifter-imagegw` package to enable easy activation of the Shifter
Image Gateway services on RHEL7.

If `--with slurm` is passed to `rpmbuild`, translate that into the
equivalent of `--define 'with_slurm %{_prefix}'`.

Distribute documentation directory so that it can be included in the
`%doc` for the `shifter` RPM.  Ditto for CLE6 Ansible roles.

Add example image gateway scripts from `contrib/` as documentation to
the `shifter-imagegw` RPM.