contents of /dev are bind mounted one by one into the container
Shifter originally bind mounted the entire /dev into the container.
The problem with this approach was that any change applied to the
bind-mounted /dev would propagate outside the container. Now
the contents of /dev within the container can be unmounted and
removed without having any effect in the /dev of the host. E.g.
the unused NVIDIA devices can be unbinded and removed to prevent the
container to use them.