mpi support: added inspection of container's ld.so.cache
The script activate_mpi_support.sh searches the ls.so.cache of
the container for references to MPI libraries. When a reference
is found, the corresponding site MPI library is bind mounted into
the container to override the container's library. If no
ABI-compatible site library is found, an error is generated.