added function shifter_setupenv_site_resources
This function recursively searches the site resources
folder. Note: the location of the site resources folder
is specified by the system administrator in the configuration file.

If an executable file is found in the site resources folder (or any
subfolder), the parent directory of such file is added to the
environment variable PATH.

If a shared library is found in the site resources folder (or any
subfolder), the parent directory of such library is added to the
environment variable LD_LIBRARY_PATH.