mpi support: check that container has MPI libraries
Now when the user requires MPI support (--mpi command line option),
activate_mpi_support.sh checks that the container's dynamic linker
is configured to load MPI libraries (check output of 'ldconfig -p').
If the dynamic linker is not configured as expected an error message
is issued and the program aborts. Thus, at least in some cases,
we are able to enforce the user to properly configure the dynamic linker
instead of doing LD_LIBRARY_PATH hacks in the container.