gpu support: adapted to only use siteResources
The GPU support logic used to use and pass around three variables contaning
the paths of the container's GPU support binaries, libraries and 64-bit libraries.
Now it is no longer the case. We only use the variable siteResources, which is also
used in similar ways for the MPI support part of Shifter.