MPI support script: fixed ABI compatibiliy check
The minor version number of the MPI libraries used to be ignored,
i.e. not checked for compatibility, when the site's library version
string didn't contain a minor number. E.g. if the host/site
had an MPI library libmpi.so.12 then a container's library
libmpi.so.12.5 would have been considered ABI compatible, which
is wrong. This commit fixes the problem. Now, if the version
string of the MPI library doesn't contain version numbers, we
assume them to be zero.