WLM integration: always activate GPU support
Always attempt to automatically enable GPU support. We decided
to avoid the alternative solution of having the user to explicitely
request the GPU support activation through an extra workload manager's
option. Such alternative solution wouldn't have been consistent with
the interface of shifter when used without workload manager
integration as well as the interface of nvidia-docker.