Fix for Issue 244
Imported images have an extra './' in the tar file paths.  If a later
layer that builds on that import changes a directory to a file or link,
the paths will be different.  So this normalizes the paths across all the
layers by stripping off any './' prefix.

It adds a test too.