Convert async workethreads to use classes
This refactors the worker thread class and image request class.
Now there is one general submit function in the worker thread
class.  The ImageRequest class has been replaced with a set
of AsyncRequests classes for each operation.  This makes
the code a bit cleaner and the inputs for each async tasks
are more clearly defined.